+++ sdk/messaging/azservicebus/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release History
 
+## 1.9.1 (TBD)
+
+### Bugs Fixed
+
+- Receiver's, in ReceiveModeReceiveAndDelete, now allow ReceiveMessages() calls after Receiver.Close. These calls will only draw from any internally cached messages that 
+  accumulated between the final call to ReceiveMessages() and Close. See the example "ExampleReceiver_ReceiveMessages_receiveAndDelete", in [example_receiver_test.go](./example_receiver_test.go)
+  for an example. (PR#TBD)
+
 ## 1.9.0 (2025-05-06)
 
 ### Features Added